routes/requirements.txt
fastapi
uvicorn
sqlalchemy
databases
pydantic
bcrypt
python-jose
passlib
sklearn
pandas
numpy
matplotlib
typing
Pillow
